Add versioning to PiCli FaaS calls
This change allows for the user to specify version numbers for
each pipe and for the policy defined in the validator pipe.
This allows for granular control of the pipeline during periods
of heavy change to the Piedpiper services.

This change required modifications to the piedpiper-validator-faas
and an additional git repository. Please see the documentation
for piedpiper-validator-faas for more information on those changes.

In each pipe_vars.d file there is now a required `version` field
which expects a version number (symver) or `latest`. This correlates
to an OpenFaaS function with the same version. For example, the
pipe_vars.d/pi_lint.yml version of `0.0.1` will hit a url of
{self.config.endpoint}/piedpiper-{self.name}-function-0-0-1. When
targetting `latest` it will leave off the version number from the URL.

Additionally `pipe_vars.d/pi_validate.yml` has a new required version
field under the `policy` dict. This will be passed to the validator
FaaS in the run_vars and will be used to control which version of
the policy will be applied when running the validator pipe.

Example configuration files can be found under tests/functional