Add SAST command
This commits adds a SAST command to PiCli. This is implemented
in much the same way as the linter command. PiCli will parse
its group_vars and file_vars for a pi_sast dictionary key and
add the files specified to the run configuration. Currently
this has just been tested with a noop SAST analyzer.