Add style command
This commit adds the 'style' command in-place of the 'lint
command. The intention is to divide the functionalities of a pure
'style' tool and a 'sast' tool, and put then in a 'lint' container.

The word 'linter' has been replaced by 'styler' in the code. A new
command called 'style' has been created which runs code-style tests.
The 'lint' command will now run the sequence ['style', 'sast'] when
called.

Splitting this up allows us to define more granular control over
our commands.