Refactor packaging
Migrate to setuptools-scm and setup.py. This follows on the back
of https://github.com/ansible/molecule/pull/1895. There are so
many different standards for specifying Python dependencies
so we are just going to pick one and stick with it.
All runtime dependencies should be specified in setup.cfg.
All test dependencies should be specified in tox.ini. Since
we are using Tox as our test execution environment we can simply
pin dependencies here.